+++ src/app/app.routes.ts
@@ -1,6 +1,6 @@
 import { Routes } from "@angular/router"
 
-import { WelcomePage } from "./pages/welcome/welcome"
+import { WelcomePage } from "./pages/welcome/WelcomePage"
 import { ProsodyPage } from "./pages/new-poem/step-1/ProsodyPage"
 import { MivesPage } from "./pages/new-poem/step-2/MivesPage"
 import { WeightsPage } from "./pages/new-poem/step-3/WeightsPage"
